onChangeMap event, with corrections
